pyproject.toml
[project]
name = "photo-fix"
version = "0.1.0"
description = "CLI tool to compare photo folders and identify missing files"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "imagehash>=4.3.0",
    "Pillow>=10.0.0",
    "pillow-heif>=0.16.0",
    "python-dateutil>=2.8.0",
    "nicegui>=2.0.0",
]

[project.scripts]
photo-restore = "photo_restore:main"
